libLLVM: Add ExecutionEngine and JIT for CrOS llvmpipe
Porting ag/1716203 to AOSP.

Chrome OS uses software rendering for Android graphics when running in a
VM, specifically Mesa's llvmpipe driver.
llvmpipe needs ExecutionEngine, MCJIT, OrcJIT and RuntimeDyld components
to be added to the LLVM runtime libLLVM.so.
This CL adds the necessary components. It increases the size of
libLLVM.so by ~360KB, from 18.9MB to 19.2MB for dragon.

Bug: 62387593
Test: mmm external/llvm. Check that libLLVMMCJIT.a, libLLVMOrcJIT.a,
libLLVMExecutionEngine.a and libLLVMRuntimeDyld.a intermediates are
built and included in libLLVM.so by spot-checking symbols.

Change-Id: I5c54535f0768daced221514582c550be1f3a8e89